move test_app out of integration_tests
