+++ flutter_theoplayer_sdk/example/integration_test/playback_test.dart
@@ -11,7 +11,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 import 'package:theoplayer/theoplayer.dart';
 
-import 'test_app.dart';
+import '../integration_test_app/test_app.dart';
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
