+++ flutter_theoplayer_sdk/example/integration_test/player_creation_test.dart
@@ -10,7 +10,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 
-import 'test_app.dart';
+import '../integration_test_app/test_app.dart';
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
